refactor(functions): move page-tracking and rarityCalc to @functions directory
This commit relocates the `page-tracking` and `rarityCalc` modules to the `@functions` directory to improve code organization and maintainability. All import paths in related components and pages have been updated to reflect this change.
