+++ src/functions/rarityCalc.ts
